requirements.txt
langchain==0.3.27
langchain_chroma==0.2.6
langchain_community==0.3.30
langchain_core==0.3.77
langchain_openai==0.3.34
langchain_text_splitters==0.3.11
langgraph==0.6.8
pandas==2.3.3
python-dotenv==1.1.1
streamlit==1.42.0
